Isolated CRUD student event functions into utils file for better organization
